feat: support upload file from an URL
